refactor: simpiify profile util with profilemanager

+++ ofscraper/utils/profiles/manage.py
@@ -4,14 +4,12 @@ import shutil
 from rich import print
 
 import ofscraper.prompts.prompts as prompts
-import ofscraper.utils.config.config as config_
 import ofscraper.utils.console as console
-import ofscraper.utils.of_env.of_env as of_env
 import ofscraper.utils.paths.common as common_paths
 import ofscraper.utils.paths.manage as manage
 import ofscraper.utils.profiles.data as profile_data
 import ofscraper.utils.profiles.tools as tools
-import ofscraper.utils.settings as settings
+import ofscraper.managers.manager as manager
 
 
 log = logging.getLogger("shared")
@@ -20,23 +18,25 @@ currentData = None
 currentProfile = None
 
 
-def change_profile():
+def change_default_profile():
     tools.print_profiles()
     profile = prompts.get_profile_prompt(profile_data.get_profile_names())
-    config_.update_config(of_env.getattr("mainProfile"), profile)
-    args = settings.get_args()
-    # remove profile argument
-    args.profile = None
-    settings.update_args(args)
-    print(f"[green]Successfully changed profile to[/green] {profile}")
+    manager.Manager.profile_manager.switch_profile(profile,permanent=True)
 
 
+def change_active_profile():
+    tools.print_profiles()
+    profile = prompts.get_profile_prompt(profile_data.get_profile_names())
+    manager.Manager.profile_manager.switch_profile(profile,permanent=False)
 def delete_profile():
     tools.print_profiles()
     profile = prompts.get_profile_prompt(profile_data.get_profile_names())
+    if profile==profile_data.get_active_profile():
+        log.error("can't delete active profile")
+        return
     p = common_paths.get_config_home() / profile
     shutil.rmtree(p)
-
+    manager.Manager.profile_manager.delete_profile()
     print(f"[green]Successfully deleted[/green] {profile}")
 
 
@@ -44,39 +44,33 @@ def create_profile():
     tools.print_profiles()
     name = tools.profile_name_fixer(prompts.create_profiles_prompt())
     manage.create_profile_path(name)
-    if prompts.change_default_profile() == "Yes":
-        config_.update_config(of_env.getattr("mainProfile"), name)
+    choice=prompts.change_selected_profile()
+    if choice=="default":
+        manager.Manager.profile_manager.switch_profile(name,permanent=True)
+    if choice=="active":
+        manager.Manager.profile_manager.switch_profile(name,permanent=False)
     console.print(
         "[green]Successfully created[/green] {dir_name}".format(dir_name=name)
     )
 
-
 def edit_profile_name():
+    """
+    Handles the user prompts for renaming a profile and calls the manager
+    to perform the action.
+    """
+    # --- 1. Get User Input ---
     tools.print_profiles()
     profiles_ = profile_data.get_profiles()
     old_profile_name = prompts.edit_profiles_prompt(profiles_)
+
+    # Exit if user cancels
+    if not old_profile_name:
+        return
+
     new_profile_name = tools.profile_name_fixer(
         prompts.new_name_edit_profiles_prompt(old_profile_name)
     )
-    configFolder = common_paths.get_config_home()
-    oldFolder = configFolder / old_profile_name
-    newFolder = configFolder / new_profile_name
-
-    if not oldFolder.exists():
-        manage.create_profile_path(new_profile_name)
-    else:
-        shutil.move(oldFolder, newFolder)
-        change_current_profile(new_profile_name, old_profile_name)
-        print(
-            f"[green]Successfully changed[green] '{old_profile_name}' to '{new_profile_name}'"
-        )
-
-
-def change_current_profile(new_profile, old_profile):
-    args = settings.get_args()
-    if args.profile == old_profile:
-        args.profile = new_profile
-        settings.update_args(args)
-    # required because name has changed
-    if old_profile == profile_data.get_current_config_profile():
-        config_.update_config(of_env.getattr("mainProfile"), new_profile)
+    # Exit if user cancels
+    if not new_profile_name:
+        return
+    manager.Manager.profile_manager.rename_profile(old_profile_name, new_profile_name)
